improved schema validation
